+++ account-kit/java/signer-sdk/src/main/java/com/alchemy/aa/core/exceptions/StamperNotInitializedException.java
@@ -2,6 +2,5 @@ package com.alchemy.aa.core.exceptions;
 
 public class StamperNotInitializedException extends RuntimeException {
 
-    public StamperNotInitializedException() {
-    }
+  public StamperNotInitializedException() {}
 }