+++ examples/ui-demo/src/components/small-cards/Button.tsx
@@ -8,10 +8,7 @@ export const Button = ({
   React.ComponentProps<"button"> & { className?: string }
 >) => {
   return (
-    <button
-      className={`border bg-bg-surface-default text-fg-primary rounded-lg px-3 text-sm font-semibold h-10 flex items-center justify-center text-center disabled:bg-bg-surface-disabled disabled:text-fg-disabled ${className}`}
-      {...rest}
-    >
+    <button className={`akui-btn-auth akui-btn ${className}`} {...rest}>
       {children}
     </button>
   );