+++ site/sidebar/reference/account-kit/signer.ts
@@ -92,6 +92,10 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
             link: "/reference/account-kit/signer/classes/BaseAlchemySigner/disconnect",
           },
           {
+            text: "experimental_toSolanaSigner",
+            link: "/reference/account-kit/signer/classes/BaseAlchemySigner/experimental_toSolanaSigner",
+          },
+          {
             text: "exportWallet",
             link: "/reference/account-kit/signer/classes/BaseAlchemySigner/exportWallet",
           },
@@ -102,6 +106,10 @@ export const accountKitSignerReferenceSidebar: SidebarItem[] = [
           {
             text: "getAuthDetails",
             link: "/reference/account-kit/signer/classes/BaseAlchemySigner/getAuthDetails",
+          },
+          {
+            text: "getConfig",
+            link: "/reference/account-kit/signer/classes/BaseAlchemySigner/getConfig",
           },
           {
             text: "getUser",